Timestep reduced
Went over it again.  Discovered that oscillations would get larger as it
ran.  Turns out that was because the timestep was too large - it would
travel past the point of zero energy during that large time step.
Reducing the timestep reduced this effect, although it's still there.
Currently at 1 ms, and oscillations are manageable, but smaller would be
better.  At less than 1 ms, I begin to worry about missing deadlines.
This could be an argument for threading the physics engine.